fix: renamed case generation to case creation
