+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseGenerationForm/components/SubmitSection/index.ts
