+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseGenerationForm/hooks/useFormSchema/useFormSchema.ts
