+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/context/case-creation-context/case-creation.context.ts
@@ -1,0 +1,9 @@
+import { CaseGenerationContext } from '@/pages/Entities/components/CaseCreation/context/case-creation-context/types';
+import { createContext } from 'react';
+
+export const initialCaseCreationContext = {
+  isMultipleCasesCreation: false,
+  isOpen: false,
+} as CaseGenerationContext;
+
+export const caseCreationContext = createContext<CaseGenerationContext>(initialCaseCreationContext);
