+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/context/case-creation-context/hocs/withCaseCreation/index.ts
@@ -1,0 +1,1 @@
+export * from './withCaseCreation';
