+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/context/case-creation-context/hooks/useCaseCreationContext/useCaseCreationContext.ts
@@ -1,0 +1,4 @@
+import { caseCreationContext } from '@/pages/Entities/components/CaseCreation/context/case-creation-context/case-creation.context';
+import { useContext } from 'react';
+
+export const useCaseCreationContext = () => useContext(caseCreationContext);
