+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/context/case-creation-context/types.ts
@@ -1,0 +1,11 @@
+export interface CaseCreationContextState {
+  isMultipleCasesCreation: boolean;
+  isOpen: boolean;
+}
+
+export interface CaseCreationApi {
+  toggleMultiCaseCreation: () => void;
+  setOpen: (open: boolean) => void;
+}
+
+export type CaseGenerationContext = CaseCreationContextState & CaseCreationApi;
