+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/hooks/useCaseCreationWorkflowDefinition/index.ts
@@ -1,0 +1,1 @@
+export * from './useCaseCreationWorkflowDefinition';
