+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/index.ts
@@ -1,0 +1,1 @@
+export * from './CaseCreation';
