--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/context/case-generation-context/case-generation.context.ts
@@ -1,11 +1,0 @@
-import { CaseGenerationContext } from '@/pages/Entities/components/CaseGeneration/context/case-generation-context/types';
-import { createContext } from 'react';
-
-export const initialCaseGenerationContext = {
-  isMultipleCasesCreation: false,
-  isOpen: false,
-} as CaseGenerationContext;
-
-export const caseGenerationContext = createContext<CaseGenerationContext>(
-  initialCaseGenerationContext,
-);
