--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/context/case-generation-context/hocs/withCaseGeneration/index.ts
@@ -1,1 +1,0 @@
-export * from './withCaseGeneration';
