--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/context/case-generation-context/hooks/useCaseGenerationContext/index.ts
@@ -1,1 +1,0 @@
-export * from './useCaseGenerationContext';
