--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/context/case-generation-context/hooks/useCaseGenerationContext/useCaseGenerationContext.ts
@@ -1,4 +1,0 @@
-import { caseGenerationContext } from '@/pages/Entities/components/CaseGeneration/context/case-generation-context/case-generation.context';
-import { useContext } from 'react';
-
-export const useCaseGenerationContext = () => useContext(caseGenerationContext);
