--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/context/case-generation-context/types.ts
@@ -1,11 +1,0 @@
-export interface CaseGenerationContextState {
-  isMultipleCasesCreation: boolean;
-  isOpen: boolean;
-}
-
-export interface CaseGenerationApi {
-  toggleMultiCaseCreation: () => void;
-  setOpen: (open: boolean) => void;
-}
-
-export type CaseGenerationContext = CaseGenerationContextState & CaseGenerationApi;
