--- apps/backoffice-v2/src/pages/Entities/components/CaseGeneration/index.ts
@@ -1,1 +1,0 @@
-export * from './CaseGeneration';
